// src/types/app.ts
export type PageType = 
  | 'landing' 
  | 'auth' 
  | 'onboarding' 
  | 'dashboard' 
  | 'inbox' 
  | 'jobSearch' 
  | 'smartMatchJobDetail'
  | 'legacyJobSearch' 
  | 'applications' 
  | 'applicationProgress'
  | 'interviews' 
  | 'upcomingInterviewDetail'
  | 'pastInterviewDetail'
  | 'jobFocusDetail'
  | 'createJobFocus'
  | 'aiCoach' 
  | 'growthCenter' 
  | 'careerProfile'
  | 'profilePreview'
  | 'settings' 
  | 'jobDetail' 
  | 'semanticFit' 
  | 'employerDashboard' 
  | 'employerHome' 
  | 'employerInbox' 
  | 'employerTalentMarketplace'
  | 'employerTalentSearch'
  | 'employerJobs' 
  | 'employerJobCreation'
  | 'employerJobDetail'
  | 'employerApplications' 
  | 'employerApplicationDetail'
  | 'employerInterviews' 
  | 'employerCreateInterview'
  | 'employerFeedbackCenter'
  | 'employerCandidates' 
  | 'employerCandidateProfile'
  | 'employerCandidatePreview'
  | 'employerAnalytics' 
  | 'employerCompany' 
  | 'employerTests' 
  | 'employerAssessments'
  | 'employerCaseStudies'
  | 'employerCaseStudyCreation'
  | 'employerSettings' 
  | 'employerBilling' 
  | 'candidateMarketplace' 
  | 'adminPanel' 
  | 'billing' 
  | 'rejectionFeedback' 
  | 'testSimulation' 
  | 'redFlagReporting' 
  | 'jobCreation' 
  | 'hrAnalytics' 
  | 'enhancedAdminPanel' 
  | 'companyCreation';

export type UserType = 'candidate' | 'employer' | 'admin';